+++ CHANGELOG.md
@@ -49,6 +49,9 @@
 - TLSRoute support: Add `TLSRoute` reconciler in hybrid gateway controllers
   and add translator of `TLSRoute`.
   [#3763](https://github.com/Kong/kong-operator/pull/3763)
+- EventGatewayDataPlane: introduces the eventgateway.konghq.com/v1alpha1 API group with
+  a new DataPlane kind that manages a keg binary Deployment connecting to Konnect
+  via a referenced KonnectEventGateway resource.
 - TLSRoute support: Configure DataPlaneOption in created `DataPlane` to
   configure Kong DataPlane deployment and ingress service for listeners with
   `TLS` protocol.